Move `IEventDecryptionResult` into `@types/crypto.ts`
