+++ src/crypto/algorithms/megolm.ts
@@ -20,6 +20,7 @@ limitations under the License.
 
 import { v4 as uuidv4 } from "uuid";
 
+import type { IEventDecryptionResult } from "../../@types/crypto";
 import { logger } from "../../logger";
 import * as olmlib from "../olmlib";
 import {
@@ -38,13 +39,7 @@ import { IOlmSessionResult } from "../olmlib";
 import { DeviceInfoMap } from "../DeviceList";
 import { IContent, MatrixEvent } from "../../models/event";
 import { EventType, MsgType, ToDeviceMessageId } from "../../@types/event";
-import {
-    IMegolmEncryptedContent,
-    IEventDecryptionResult,
-    IMegolmSessionData,
-    IncomingRoomKeyRequest,
-    IEncryptedContent,
-} from "../index";
+import { IMegolmEncryptedContent, IMegolmSessionData, IncomingRoomKeyRequest, IEncryptedContent } from "../index";
 import { RoomKeyRequestState } from "../OutgoingRoomKeyRequestManager";
 import { OlmGroupSessionExtraData } from "../../@types/crypto";
 import { MatrixError } from "../../http-api";
